JWT/requirements.txt
Flask==1.1.2
google-cloud-datastore==1.12.0
requests_oauthlib==3.1.0